Implementing Nalu Navigation
The most significant changes involve the integration of an `INavigationService` parameter into the constructors of the `App` and `AppShell` classes, and the modification of the `AppShell` class to inherit from `NaluShell` instead of `Shell`. Additionally, the `nalu:Navigation.PageType` attribute in `AppShell.xaml` has been updated, and the lambda function in `MauiProgram.cs` has been adjusted to replace `"Page"` with `"PageViewModel"` in the page type name.

List of Changes:

1. The `App` class in `App.xaml.cs` now accepts an `INavigationService` parameter in its constructor and initializes the `MainPage` with this navigation service. (Refer to `App.xaml.cs`)

2. The `nalu:Navigation.PageType` attribute of the `ShellContent` element in `AppShell.xaml` has been updated from `"MainPage"` to `"local:MainPage"`. (Refer to `AppShell.xaml`)

3. The `AppShell` class in `AppShell.xaml.cs` now inherits from `NaluShell` instead of `Shell` and accepts an `INavigationService` parameter in its constructor, which is passed to the base constructor along with the `MainPage` type. (Refer to `AppShell.xaml.cs`)

4. The lambda function passed to the `UseNaluNavigation` method in `MauiProgram.cs` has been modified to replace `"Page"` with `"PageViewModel"` in the page type name. (Refer to `MauiProgram.cs`)